Block Cluster object from being deleted before the namespace is gone (#10359)
* extend HasOnlyFinalizer() to be able to check for multiple finalizers

* add namespace-cleanup finalizer on Cluster objects

* emit events during cluster deletion

* fix tests

* fix copyright

* flip order: namespace first, then delete credentials

* do not attempt to update the cluster health during shutdowns

* introduce cleanup-addons finalizer, wait for addon cleanup during cluster deletion

* auto-update controller should not update cluster during deletions

* more events! more logging

* fix potential update() call on a gone cluster because the provider always cleans up, regardless of its finalizer

* logic is hard

* revert the bad idea of deleting addons cleanly

* clarify, previous wording could be confused with actual provider stuff (like subnets, VPCs, ..)

* fine-tune logging

* improve logging

* debugging

* debugging harder

* cleanup

* improve wording

* gimps

* PR feedback

* marvin wanted to have this already, so here we go

+++ pkg/clusterdeletion/etcdbackupconfig.go
@@ -24,6 +24,7 @@ import (
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
 	kuberneteshelper "k8c.io/kubermatic/v2/pkg/kubernetes"
 
+	corev1 "k8s.io/api/core/v1"
 	ctrlruntimeclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
@@ -32,21 +33,26 @@ func (d *Deletion) cleanupEtcdBackupConfigs(ctx context.Context, cluster *kuberm
 		return nil
 	}
 
-	// always attempt to cleanup, even if the controllers might be disabled now
-	backupConfigs := &kubermaticv1.EtcdBackupConfigList{}
-	if err := d.seedClient.List(ctx, backupConfigs, ctrlruntimeclient.InNamespace(cluster.Status.NamespaceName)); err != nil {
-		return fmt.Errorf("failed to get EtcdBackupConfigs: %w", err)
-	}
-
-	if len(backupConfigs.Items) > 0 {
-		for _, backupConfig := range backupConfigs.Items {
-			if err := d.seedClient.Delete(ctx, &backupConfig); err != nil {
-				return fmt.Errorf("failed to delete EtcdBackupConfig %q: %w", backupConfig.Name, err)
-			}
+	if cluster.Status.NamespaceName != "" {
+		// always attempt to cleanup, even if the controllers might be disabled now
+		backupConfigs := &kubermaticv1.EtcdBackupConfigList{}
+		if err := d.seedClient.List(ctx, backupConfigs, ctrlruntimeclient.InNamespace(cluster.Status.NamespaceName)); err != nil {
+			return fmt.Errorf("failed to get EtcdBackupConfigs: %w", err)
 		}
 
-		return nil
+		if len(backupConfigs.Items) > 0 {
+			for _, backupConfig := range backupConfigs.Items {
+				if err := d.seedClient.Delete(ctx, &backupConfig); err != nil {
+					return fmt.Errorf("failed to delete EtcdBackupConfig %q: %w", backupConfig.Name, err)
+				}
+			}
+
+			d.recorder.Eventf(cluster, corev1.EventTypeNormal, "EtcdBackupConfigCleanup", "There are %d EtcdBackupConfig objects waiting for deletion.", len(backupConfigs.Items))
+			return nil
+		}
 	}
+
+	d.recorder.Event(cluster, corev1.EventTypeNormal, "EtcdBackupConfigCleanup", "Cleanup has been completed.")
 
 	return kuberneteshelper.TryRemoveFinalizer(ctx, d.seedClient, cluster, apiv1.EtcdBackupConfigCleanupFinalizer)
 }
